Add sign inference on type as a last resort
Previously e%:sgn was failing when an unknown (or no, i.e., for which
no canonical instance is registered) operation was encountered in the
expression e, now some sign information can be inferred from the type
of the unknown sub expression.

This is generalizing the canonical instance already existing for the
type nat.

This is needed in https://github.com/math-comp/analysis/pull/375 to be
able to infer that `|x|%E is non negative when nothing is known on x.

lignes